feat(rebalancer): move factory outside metrics scope
